refacto: rename to expand-urls
