+++ source/content.js
@@ -63,7 +63,7 @@ function onDomReady() {
 	enableFeature(features.cleanNavbarDropdown);
 	enableFeature(features.keyboardShortcuts);
 	enableFeature(features.preserveTextMessages);
-	enableFeature(features.expandUrlsRemoveUtms);
+	enableFeature(features.expandUrls);
 
 	onRouteChange(() => {
 		enableFeature(features.autoLoadNewTweets);
@@ -82,7 +82,7 @@ function onDomReady() {
 			enableFeature(features.hidePromotedTweets);
 			enableFeature(features.renderInlineCode);
 			enableFeature(features.imageAlternatives);
-			enableFeature(features.expandUrlsRemoveUtms);
+			enableFeature(features.expandUrls);
 		});
 	});
